Added special effects logic and cooldown on special attacks

+++ src/Library/Attack.cs
@@ -4,7 +4,7 @@ public class Attack : IAttack
 {
     public string Name { get; }
     public Type Type{get; }
-    public int Accuracy {get; private set; }
+    public double Accuracy {get; private set; }
     public int Power {get; private set; }
 
     public Attack(string name, Type type, int accuracy, int power)
